refactor(ai): split ai_ingest into per-agent files
- split ai_ingest.go into claude/codex/antigravity ingest files
- decompose ingestClaudeHook into focused helpers
