+++ internal/app/ai_ingest.go
@@ -16,11 +16,9 @@ import (
 
 	"github.com/crevissepartners/projmux/internal/config"
 	"github.com/crevissepartners/projmux/internal/core/notify"
-	antigravityadapter "github.com/crevissepartners/projmux/internal/core/usage/adapters/antigravity"
 	intmux "github.com/crevissepartners/projmux/internal/integrations/mux"
 )
 
-const claudeTranscriptTailLimit = 256 * 1024
 const (
 	aiBellDedupeOption = "@projmux_ai_bell_notified_at"
 	aiBellDedupeWindow = 5 * time.Second
@@ -50,18 +48,6 @@ var aiBellPaneFormats = []string{
 
 var aiBellPaneFormat = intmux.JoinFormats(intmux.FieldDelimiter, aiBellPaneFormats...)
 
-type codexHookPayload struct {
-	EventName      string
-	ThreadID       string
-	SessionID      string
-	TurnID         string
-	CWD            string
-	TranscriptPath string
-	Model          string
-	ToolName       string
-	ToolInput      map[string]any
-}
-
 type aiPaneMatchInput struct {
 	CWD       string
 	ThreadID  string
@@ -86,41 +72,6 @@ type aiIngestLogEntry struct {
 	ThreadID  string `json:"thread_id,omitempty"`
 	SessionID string `json:"session_id,omitempty"`
 	TurnID    string `json:"turn_id,omitempty"`
-}
-
-type claudeHookPayload struct {
-	EventName        string
-	SessionID        string
-	CWD              string
-	TranscriptPath   string
-	NotificationType string
-	Message          string
-	Prompt           string
-	ToolName         string
-	ToolUseID        string
-	ToolInput        map[string]any
-	ErrorType        string
-	ErrorMessage     string
-	SubagentType     string
-	SubagentID       string
-	TeammateName     string
-	TeammateID       string
-	TeammateContext  string
-}
-
-type antigravityHookPayload struct {
-	EventName                  string
-	CWD                        string
-	ConversationID             string
-	TranscriptPath             string
-	TerminationReason          string
-	Error                      string
-	FullyIdle                  bool
-	FullyIdleSet               bool
-	ToolConfirmationPending    bool
-	ToolConfirmationPendingSet bool
-	AgentState                 string
-	ContextWindow              string
 }
 
 func (c *aiCommand) runIngest(args []string, stdout, stderr io.Writer) error {
@@ -399,554 +350,6 @@ func composeBellNotifyText(info bellPaneInfo) string {
 	return "bell · " + context
 }
 
-func (c *aiCommand) ingestCodexHook(data []byte) error {
-	payload, err := parseCodexHookPayload(data)
-	if err != nil {
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Result: "error", Reason: err.Error()})
-		return err
-	}
-
-	paneID := c.matchAIPane(aiPaneMatchInput{
-		CWD:       payload.CWD,
-		ThreadID:  payload.matchThreadID(),
-		SessionID: payload.SessionID,
-	})
-	if paneID == "" {
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "ignored", Reason: "no matching pane", CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-		return nil
-	}
-
-	metadata := payload.codexHookMetadata()
-	action := c.aiHookEffectiveAction(aiHookProviderCodex, payload.EventName)
-	switch payload.EventName {
-	case "UserPromptSubmit":
-		if action.Action == aiHookActionQuiet {
-			c.quietCodexHook(paneID, payload, aiHookQuietReason(action))
-			return nil
-		}
-		c.markAIHookPane(paneID, aiModeCodex, payload.CWD, payload.matchThreadID(), payload.SessionID, "")
-		if err := c.applyAIStatusWithNotify("thinking", paneID, attentionNotifyInput{
-			Metadata:  metadata,
-			BadgeKind: aiBadgeKindInProgress,
-		}); err != nil {
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-			return err
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "state", Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-		return nil
-	case "Stop":
-		if action.Action == aiHookActionQuiet {
-			c.quietCodexHook(paneID, payload, aiHookQuietReason(action))
-			return nil
-		}
-		c.markAIHookPane(paneID, aiModeCodex, payload.CWD, payload.matchThreadID(), payload.SessionID, "")
-		body := formatCodexHookStopNotifyBody(payload)
-		if action.Action == aiHookActionState {
-			if err := c.applyAIStatusStateOnly("waiting", paneID, attentionNotifyInput{
-				ID:        codexHookNotifyID(payload, "stop"),
-				Text:      body.Text,
-				Severity:  body.Severity,
-				Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-				Force:     true,
-				BadgeKind: aiBadgeKindResponseComplete,
-			}); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "state", Reason: aiHookStateReason(action), Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-			return nil
-		}
-		if err := c.applyAIStatusWithNotify("waiting", paneID, attentionNotifyInput{
-			ID:        codexHookNotifyID(payload, "stop"),
-			Text:      body.Text,
-			Severity:  body.Severity,
-			Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-			Force:     true,
-			BadgeKind: aiBadgeKindResponseComplete,
-		}); err != nil {
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-			return err
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "notify", Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-		return nil
-	case "PermissionRequest":
-		if action.Action == aiHookActionQuiet {
-			c.quietCodexHook(paneID, payload, aiHookQuietReason(action))
-			return nil
-		}
-		c.markAIHookPane(paneID, aiModeCodex, payload.CWD, payload.matchThreadID(), payload.SessionID, "")
-		body := formatCodexHookPermissionNotifyBody(payload)
-		if action.Action == aiHookActionState {
-			if err := c.applyAIStatusStateOnly("waiting", paneID, attentionNotifyInput{
-				ID:        codexHookNotifyID(payload, "permission"),
-				Text:      body.Text,
-				Severity:  body.Severity,
-				Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-				Force:     true,
-				BadgeKind: aiBadgeKindApprovalRequired,
-			}); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "state", Reason: aiHookStateReason(action), Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-			return nil
-		}
-		if err := c.applyAIStatusWithNotify("waiting", paneID, attentionNotifyInput{
-			ID:        codexHookNotifyID(payload, "permission"),
-			Text:      body.Text,
-			Severity:  body.Severity,
-			Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-			Force:     true,
-			BadgeKind: aiBadgeKindApprovalRequired,
-		}); err != nil {
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-			return err
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "notify", Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-		return nil
-	case "PreToolUse", "PostToolUse", "PreCompact", "PostCompact", "SessionStart":
-		if c.shouldPushGenericCodexHookNotify(action) {
-			if err := c.pushGenericCodexHookNotify(paneID, payload); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "notify", Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-			return nil
-		}
-		c.quietCodexHook(paneID, payload, aiHookNoHandlerReason(action))
-		return nil
-	default:
-		c.quietCodexHook(paneID, payload, aiHookNoHandlerReason(action))
-		return nil
-	}
-}
-
-func (c *aiCommand) ingestClaudeHook(data []byte) error {
-	payload, err := parseClaudeHookPayload(data)
-	if err != nil {
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Result: "error", Reason: err.Error()})
-		return err
-	}
-
-	paneID := c.matchAIPane(aiPaneMatchInput{
-		CWD:       payload.CWD,
-		SessionID: payload.SessionID,
-	})
-	if paneID == "" {
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "ignored", Reason: "no matching pane", CWD: payload.CWD, SessionID: payload.SessionID})
-		return nil
-	}
-
-	c.markAIHookPane(paneID, aiModeClaude, payload.CWD, "", payload.SessionID, payload.TranscriptPath)
-	metadata := payload.claudeMetadata()
-	action := c.aiHookEffectiveAction(aiHookProviderClaude, payload.EventName)
-
-	switch payload.EventName {
-	case "UserPromptSubmit":
-		if action.Action == aiHookActionQuiet {
-			c.quietClaudeHook(paneID, payload, aiHookQuietReason(action))
-			return nil
-		}
-		if err := c.applyAIStatusWithNotify("thinking", paneID, attentionNotifyInput{
-			Metadata:  metadata,
-			BadgeKind: aiBadgeKindInProgress,
-		}); err != nil {
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return err
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "state", Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-		return nil
-	case "Notification":
-		if action.Action == aiHookActionQuiet {
-			c.quietClaudeHook(paneID, payload, aiHookQuietReason(action))
-			return nil
-		}
-		body := formatClaudeNotificationNotifyBody(payload)
-		badgeKind := aiBadgeKindForNotifyCategory(body.Category)
-		if action.Action == aiHookActionState {
-			if err := c.applyAIStatusStateOnly("waiting", paneID, attentionNotifyInput{
-				ID:        claudeNotifyID(payload),
-				Text:      body.Text,
-				Severity:  body.Severity,
-				Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-				Force:     true,
-				BadgeKind: badgeKind,
-			}); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "state", Reason: aiHookStateReason(action), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return nil
-		}
-		if err := c.applyAIStatusWithNotify("waiting", paneID, attentionNotifyInput{
-			ID:        claudeNotifyID(payload),
-			Text:      body.Text,
-			Severity:  body.Severity,
-			Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-			Force:     true,
-			BadgeKind: badgeKind,
-		}); err != nil {
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return err
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "notify", Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-		return nil
-	case "PermissionRequest":
-		if action.Action == aiHookActionQuiet {
-			c.quietClaudeHook(paneID, payload, aiHookQuietReason(action))
-			return nil
-		}
-		body := formatClaudePermissionNotifyBody(payload)
-		if action.Action == aiHookActionState {
-			if err := c.applyAIStatusStateOnly("waiting", paneID, attentionNotifyInput{
-				ID:        claudePermissionNotifyID(payload),
-				Text:      body.Text,
-				Severity:  body.Severity,
-				Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-				Force:     true,
-				BadgeKind: aiBadgeKindApprovalRequired,
-			}); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "state", Reason: aiHookStateReason(action), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return nil
-		}
-		if err := c.applyAIStatusWithNotify("waiting", paneID, attentionNotifyInput{
-			ID:        claudePermissionNotifyID(payload),
-			Text:      body.Text,
-			Severity:  body.Severity,
-			Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-			Force:     true,
-			BadgeKind: aiBadgeKindApprovalRequired,
-		}); err != nil {
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return err
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "notify", Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-		return nil
-	case "Stop":
-		if action.Action == aiHookActionQuiet {
-			c.quietClaudeHook(paneID, payload, aiHookQuietReason(action))
-			return nil
-		}
-		message := readClaudeTranscriptLastAssistantText(payload.TranscriptPath)
-		body := formatClaudeStopNotifyBody(message)
-		if action.Action == aiHookActionState {
-			if err := c.applyAIStatusStateOnly("waiting", paneID, attentionNotifyInput{
-				ID:        claudeStopNotifyID(payload),
-				Text:      body.Text,
-				Severity:  body.Severity,
-				Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-				Force:     true,
-				BadgeKind: aiBadgeKindResponseComplete,
-			}); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "state", Reason: aiHookStateReason(action), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return nil
-		}
-		if err := c.applyAIStatusWithNotify("waiting", paneID, attentionNotifyInput{
-			ID:        claudeStopNotifyID(payload),
-			Text:      body.Text,
-			Severity:  body.Severity,
-			Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-			Force:     true,
-			BadgeKind: aiBadgeKindResponseComplete,
-		}); err != nil {
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return err
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "notify", Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-		return nil
-	case "StopFailure":
-		if action.Action == aiHookActionQuiet {
-			c.quietClaudeHook(paneID, payload, aiHookQuietReason(action))
-			return nil
-		}
-		body := formatClaudeStopFailureNotifyBody(payload)
-		if action.Action == aiHookActionState {
-			if err := c.applyAIStatusStateOnly("waiting", paneID, attentionNotifyInput{
-				ID:       claudeExtraNotifyID(payload, "stop-failure", payload.ErrorType, payload.ErrorMessage),
-				Text:     body.Text,
-				Severity: body.Severity,
-				Metadata: mergeAINotifyBodyMetadata(metadata, body),
-				Force:    true,
-			}); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "state", Reason: aiHookStateReason(action), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return nil
-		}
-		if err := c.applyAIStatusWithNotify("waiting", paneID, attentionNotifyInput{
-			ID:       claudeExtraNotifyID(payload, "stop-failure", payload.ErrorType, payload.ErrorMessage),
-			Text:     body.Text,
-			Severity: body.Severity,
-			Metadata: mergeAINotifyBodyMetadata(metadata, body),
-			Force:    true,
-		}); err != nil {
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return err
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "notify", Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-		return nil
-	case "SubagentStop":
-		if action.Action == aiHookActionNotify {
-			body := formatClaudeSubagentStopNotifyBody(payload)
-			if err := c.applyAIStatusWithNotify("waiting", paneID, attentionNotifyInput{
-				ID:       claudeExtraNotifyID(payload, "subagent-stop", payload.SubagentType, payload.SubagentID),
-				Text:     body.Text,
-				Severity: body.Severity,
-				Metadata: mergeAINotifyBodyMetadata(metadata, body),
-				Force:    true,
-			}); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "notify", Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return nil
-		}
-		if action.Action == aiHookActionState {
-			if err := c.applyAIStatusStateOnly("waiting", paneID, attentionNotifyInput{
-				ID:       claudeExtraNotifyID(payload, "subagent-stop", payload.SubagentType, payload.SubagentID),
-				Text:     formatClaudeSubagentStopNotifyBody(payload).Text,
-				Severity: notify.SeverityInfo,
-				Metadata: mergeAINotifyBodyMetadata(metadata, formatClaudeSubagentStopNotifyBody(payload)),
-				Force:    true,
-			}); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "state", Reason: aiHookStateReason(action), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return nil
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "quiet", Reason: "high-volume event", Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-		return nil
-	case "PreToolUse", "PostToolUse", "PostToolUseFailure", "PostToolBatch", "PermissionDenied", "UserPromptExpansion", "SessionStart", "SubagentStart", "PreCompact", "PostCompact", "SessionEnd", "Setup", "TaskCreated", "TaskCompleted", "Elicitation", "ElicitationResult", "ConfigChange", "InstructionsLoaded", "WorktreeCreate", "WorktreeRemove", "CwdChanged", "FileChanged":
-		c.quietClaudeHook(paneID, payload, aiHookNoHandlerReason(action))
-		return nil
-	case "TeammateIdle":
-		if action.Action == aiHookActionQuiet {
-			c.quietClaudeHook(paneID, payload, aiHookQuietReason(action))
-			return nil
-		}
-		body := formatClaudeTeammateIdleNotifyBody(payload)
-		if action.Action == aiHookActionState {
-			if err := c.applyAIStatusStateOnly("waiting", paneID, attentionNotifyInput{
-				ID:        claudeExtraNotifyID(payload, "teammate-idle", payload.TeammateName, payload.TeammateID, payload.TeammateContext),
-				Text:      body.Text,
-				Severity:  body.Severity,
-				Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-				Force:     true,
-				BadgeKind: aiBadgeKindResponseComplete,
-			}); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "state", Reason: aiHookStateReason(action), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return nil
-		}
-		if err := c.applyAIStatusWithNotify("waiting", paneID, attentionNotifyInput{
-			ID:        claudeExtraNotifyID(payload, "teammate-idle", payload.TeammateName, payload.TeammateID, payload.TeammateContext),
-			Text:      body.Text,
-			Severity:  body.Severity,
-			Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-			Force:     true,
-			BadgeKind: aiBadgeKindResponseComplete,
-		}); err != nil {
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-			return err
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "notify", Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-		return nil
-	default:
-		c.quietClaudeHook(paneID, payload, aiHookNoHandlerReason(action))
-		return nil
-	}
-}
-
-func (c *aiCommand) ingestAntigravityHook(data []byte) error {
-	payload, err := parseAntigravityHookPayload(data)
-	if err != nil {
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "antigravity-hook", Result: "error", Reason: err.Error()})
-		return err
-	}
-
-	paneID := c.matchAIPane(aiPaneMatchInput{
-		CWD:      payload.CWD,
-		ThreadID: payload.ConversationID,
-	})
-	if paneID == "" {
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "antigravity-hook", Event: payload.EventName, Result: "ignored", Reason: "no matching pane", CWD: payload.CWD, ThreadID: payload.ConversationID})
-		return nil
-	}
-
-	c.markAIHookPane(paneID, aiModeAntigravity, payload.CWD, payload.ConversationID, payload.ConversationID, payload.TranscriptPath)
-	c.persistAntigravityContextUsage(payload)
-	metadata := payload.antigravityMetadata()
-	action := c.aiHookEffectiveAction(aiHookProviderAntigravity, payload.EventName)
-
-	switch payload.EventName {
-	case "Stop":
-		if action.Action == aiHookActionQuiet {
-			c.quietAntigravityHook(paneID, payload, aiHookQuietReason(action))
-			return nil
-		}
-		body := formatAntigravityStopNotifyBody(payload)
-		if action.Action == aiHookActionState {
-			if err := c.applyAIStatusStateOnly("waiting", paneID, attentionNotifyInput{
-				ID:        antigravityNotifyID(payload, "stop"),
-				Text:      body.Text,
-				Severity:  body.Severity,
-				Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-				Force:     true,
-				BadgeKind: aiBadgeKindForNotifyCategory(body.Category),
-			}); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "antigravity-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, ThreadID: payload.ConversationID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "antigravity-hook", Event: payload.EventName, Result: "state", Reason: aiHookStateReason(action), Pane: paneID, CWD: payload.CWD, ThreadID: payload.ConversationID})
-			return nil
-		}
-		if err := c.applyAIStatusWithNotify("waiting", paneID, attentionNotifyInput{
-			ID:        antigravityNotifyID(payload, "stop"),
-			Text:      body.Text,
-			Severity:  body.Severity,
-			Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-			Force:     true,
-			BadgeKind: aiBadgeKindForNotifyCategory(body.Category),
-		}); err != nil {
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "antigravity-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, ThreadID: payload.ConversationID})
-			return err
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "antigravity-hook", Event: payload.EventName, Result: "notify", Pane: paneID, CWD: payload.CWD, ThreadID: payload.ConversationID})
-		return nil
-	case "Statusline":
-		if !payload.ToolConfirmationPending {
-			c.quietAntigravityHook(paneID, payload, "statusline has no pending tool confirmation")
-			return nil
-		}
-		if action.Action == aiHookActionQuiet {
-			c.quietAntigravityHook(paneID, payload, aiHookQuietReason(action))
-			return nil
-		}
-		body := formatAntigravityApprovalNotifyBody(payload)
-		if action.Action == aiHookActionState {
-			if err := c.applyAIStatusStateOnly("waiting", paneID, attentionNotifyInput{
-				ID:        antigravityNotifyID(payload, "approval"),
-				Text:      body.Text,
-				Severity:  body.Severity,
-				Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-				Force:     true,
-				BadgeKind: aiBadgeKindApprovalRequired,
-			}); err != nil {
-				c.appendAIIngestLog(aiIngestLogEntry{Source: "antigravity-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, ThreadID: payload.ConversationID})
-				return err
-			}
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "antigravity-hook", Event: payload.EventName, Result: "state", Reason: aiHookStateReason(action), Pane: paneID, CWD: payload.CWD, ThreadID: payload.ConversationID})
-			return nil
-		}
-		if err := c.applyAIStatusWithNotify("waiting", paneID, attentionNotifyInput{
-			ID:        antigravityNotifyID(payload, "approval"),
-			Text:      body.Text,
-			Severity:  body.Severity,
-			Metadata:  mergeAINotifyBodyMetadata(metadata, body),
-			Force:     true,
-			BadgeKind: aiBadgeKindApprovalRequired,
-		}); err != nil {
-			c.appendAIIngestLog(aiIngestLogEntry{Source: "antigravity-hook", Event: payload.EventName, Result: "error", Reason: err.Error(), Pane: paneID, CWD: payload.CWD, ThreadID: payload.ConversationID})
-			return err
-		}
-		c.appendAIIngestLog(aiIngestLogEntry{Source: "antigravity-hook", Event: payload.EventName, Result: "notify", Pane: paneID, CWD: payload.CWD, ThreadID: payload.ConversationID})
-		return nil
-	case "PostInvocation":
-		c.quietAntigravityHook(paneID, payload, aiHookNoHandlerReason(action))
-		return nil
-	default:
-		c.quietAntigravityHook(paneID, payload, aiHookNoHandlerReason(action))
-		return nil
-	}
-}
-
-// persistAntigravityContextUsage records the latest context-window
-// percentage carried by an antigravity hook into the usage state
-// directory so the usage adapter (and thus the HUD/status bar) can surface
-// it. Best-effort: antigravity exposes no 5h/weekly quota, so this
-// context-window gauge is the only usage-shaped signal available. A
-// missing or unparseable value, or a write failure, is silently ignored —
-// usage is a non-critical side channel of hook ingest.
-func (c *aiCommand) persistAntigravityContextUsage(payload antigravityHookPayload) {
-	pct, ok := antigravityadapter.ParsePercent(payload.ContextWindow)
-	if !ok {
-		return
-	}
-	baseDir, err := c.usageStateDir()
-	if err != nil {
-		return
-	}
-	_ = antigravityadapter.WriteContext(baseDir, antigravityadapter.ContextRecord{
-		Pct:       pct,
-		UpdatedAt: c.now().UTC(),
-	})
-}
-
-// usageStateDir resolves the directory the usage snapshot cache and the
-// antigravity context sidecar live in. It mirrors usageCommand.resolveStateDir
-// so the ingest writer and the adapter reader agree even when
-// PROJMUX_USAGE_STATE_DIR redirects the cache to a synced location.
-func (c *aiCommand) usageStateDir() (string, error) {
-	if override := strings.TrimSpace(c.env(stateDirEnvVar)); override != "" {
-		return override, nil
-	}
-	homeDir, err := c.homeDir()
-	if err != nil {
-		return "", fmt.Errorf("resolve home directory: %w", err)
-	}
-	paths, err := config.Homes{
-		HomeDir:    homeDir,
-		ConfigHome: c.env("XDG_CONFIG_HOME"),
-		StateHome:  c.env("XDG_STATE_HOME"),
-	}.Paths()
-	if err != nil {
-		return "", err
-	}
-	return filepath.Join(paths.StateDir, "usage"), nil
-}
-
-func (c *aiCommand) quietCodexHook(paneID string, payload codexHookPayload, reason string) {
-	c.markAIHookPane(paneID, aiModeCodex, payload.CWD, payload.matchThreadID(), payload.SessionID, "")
-	c.appendAIIngestLog(aiIngestLogEntry{Source: "codex-hook", Event: payload.EventName, Result: "quiet", Reason: reason, Pane: paneID, CWD: payload.CWD, ThreadID: payload.matchThreadID(), SessionID: payload.SessionID, TurnID: payload.TurnID})
-}
-
-func (c *aiCommand) shouldPushGenericCodexHookNotify(action aiHookActionResolution) bool {
-	return action.Action == aiHookActionNotify && action.Source == aiHookActionSourceRuntime
-}
-
-func (c *aiCommand) pushGenericCodexHookNotify(paneID string, payload codexHookPayload) error {
-	c.markAIHookPane(paneID, aiModeCodex, payload.CWD, payload.matchThreadID(), payload.SessionID, "")
-	body := formatCodexGenericHookNotifyBody(payload)
-	return c.applyAIStatusQueueOnly("waiting", paneID, attentionNotifyInput{
-		ID:            codexHookNotifyID(payload, "generic"),
-		Text:          body.Text,
-		Severity:      body.Severity,
-		Metadata:      mergeAINotifyBodyMetadata(payload.codexGenericHookMetadata(), body),
-		Force:         true,
-		SuppressHooks: true,
-	})
-}
-
-func (c *aiCommand) quietClaudeHook(paneID string, payload claudeHookPayload, reason string) {
-	c.markAIHookPane(paneID, aiModeClaude, payload.CWD, "", payload.SessionID, payload.TranscriptPath)
-	c.appendAIIngestLog(aiIngestLogEntry{Source: "claude-hook", Event: payload.EventName, Result: "quiet", Reason: reason, Pane: paneID, CWD: payload.CWD, SessionID: payload.SessionID})
-}
-
-func (c *aiCommand) quietAntigravityHook(paneID string, payload antigravityHookPayload, reason string) {
-	c.markAIHookPane(paneID, aiModeAntigravity, payload.CWD, payload.ConversationID, payload.ConversationID, payload.TranscriptPath)
-	c.appendAIIngestLog(aiIngestLogEntry{Source: "antigravity-hook", Event: payload.EventName, Result: "quiet", Reason: reason, Pane: paneID, CWD: payload.CWD, ThreadID: payload.ConversationID})
-}
-
 func (c *aiCommand) aiHookFallbackReason(provider, event string) string {
 	action, ok, err := c.aiHookCatalogAction(provider, event)
 	if err != nil {
@@ -1068,299 +471,6 @@ func formatAIIngestLogEntry(entry aiIngestLogEntry) string {
 	return strings.Join(parts, " ")
 }
 
-func parseCodexHookPayload(data []byte) (codexHookPayload, error) {
-	var raw map[string]any
-	if err := json.Unmarshal(data, &raw); err != nil {
-		return codexHookPayload{}, fmt.Errorf("parse codex hook payload: %w", err)
-	}
-	payload := codexHookPayload{
-		EventName:      firstString(raw, "hook_event_name", "event_name"),
-		ThreadID:       firstString(raw, "thread_id", "thread-id"),
-		SessionID:      firstString(raw, "session_id", "session-id"),
-		TurnID:         firstString(raw, "turn_id", "turn-id"),
-		CWD:            firstString(raw, "cwd", "workspace", "project_dir"),
-		TranscriptPath: firstString(raw, "transcript_path", "transcriptPath"),
-		Model:          firstString(raw, "model"),
-		ToolName:       firstString(raw, "tool_name", "toolName"),
-	}
-	if payload.CWD == "" {
-		payload.CWD = firstNestedString(raw["workspace"], "cwd", "path")
-	}
-	if payload.ToolName == "" {
-		payload.ToolName = firstNestedString(raw["tool"], "name", "tool_name")
-	}
-	payload.ToolInput = mapFromAny(raw["tool_input"])
-	if len(payload.ToolInput) == 0 {
-		payload.ToolInput = mapFromAny(raw["input"])
-	}
-	if len(payload.ToolInput) == 0 {
-		payload.ToolInput = mapFromAny(raw["tool"])
-		delete(payload.ToolInput, "name")
-		delete(payload.ToolInput, "tool_name")
-	}
-	return payload, nil
-}
-
-func parseClaudeHookPayload(data []byte) (claudeHookPayload, error) {
-	var raw map[string]any
-	if err := json.Unmarshal(data, &raw); err != nil {
-		return claudeHookPayload{}, fmt.Errorf("parse claude hook payload: %w", err)
-	}
-	payload := claudeHookPayload{
-		EventName:        firstString(raw, "hook_event_name", "event_name"),
-		SessionID:        firstString(raw, "session_id", "session-id"),
-		CWD:              firstString(raw, "cwd", "workspace", "project_dir"),
-		TranscriptPath:   firstString(raw, "transcript_path", "transcriptPath"),
-		NotificationType: firstString(raw, "notification_type", "notificationType"),
-		Message:          firstString(raw, "message", "text"),
-		Prompt:           firstString(raw, "prompt", "user_prompt"),
-		ToolName:         firstString(raw, "tool_name", "toolName"),
-		ToolUseID:        firstString(raw, "tool_use_id", "toolUseID", "id"),
-		ErrorType:        firstString(raw, "error_type", "errorType", "failure_type", "failureType"),
-		ErrorMessage:     firstString(raw, "error_message", "errorMessage", "message", "reason"),
-		SubagentType:     firstString(raw, "subagent_type", "subagentType", "agent_type", "agentType"),
-		SubagentID:       firstString(raw, "subagent_id", "subagentId", "agent_id", "agentId"),
-		TeammateName:     firstString(raw, "teammate_name", "teammateName", "teammate"),
-		TeammateID:       firstString(raw, "teammate_id", "teammateId"),
-		TeammateContext:  firstString(raw, "teammate_context", "teammateContext", "context", "reason", "message"),
-	}
-	if payload.CWD == "" {
-		payload.CWD = firstNestedString(raw["workspace"], "cwd", "path")
-	}
-	if payload.Message == "" {
-		payload.Message = firstNestedString(raw["notification"], "message", "text")
-	}
-	if payload.NotificationType == "" {
-		payload.NotificationType = firstNestedString(raw["notification"], "notification_type", "type")
-	}
-	if payload.ToolName == "" {
-		payload.ToolName = firstNestedString(raw["tool"], "name", "tool_name")
-	}
-	if payload.ToolUseID == "" {
-		payload.ToolUseID = firstNestedString(raw["tool"], "id", "tool_use_id")
-	}
-	if payload.ErrorType == "" {
-		payload.ErrorType = firstNestedString(raw["error"], "type", "name", "code")
-	}
-	if payload.ErrorMessage == "" {
-		payload.ErrorMessage = firstNestedString(raw["error"], "message", "text", "reason")
-	}
-	if payload.SubagentType == "" {
-		payload.SubagentType = firstNestedString(raw["subagent"], "type", "name", "kind")
-	}
-	if payload.SubagentID == "" {
-		payload.SubagentID = firstNestedString(raw["subagent"], "id", "subagent_id", "agent_id")
-	}
-	if payload.TeammateName == "" {
-		payload.TeammateName = firstNestedString(raw["teammate"], "name", "type", "kind")
-	}
-	if payload.TeammateID == "" {
-		payload.TeammateID = firstNestedString(raw["teammate"], "id", "teammate_id")
-	}
-	if payload.TeammateContext == "" {
-		payload.TeammateContext = firstNestedString(raw["teammate"], "context", "status", "reason", "message")
-	}
-	payload.ToolInput = mapFromAny(raw["tool_input"])
-	if len(payload.ToolInput) == 0 {
-		payload.ToolInput = mapFromAny(raw["input"])
-	}
-	if len(payload.ToolInput) == 0 {
-		payload.ToolInput = mapFromAny(raw["tool"])
-		delete(payload.ToolInput, "name")
-		delete(payload.ToolInput, "tool_name")
-		delete(payload.ToolInput, "id")
-		delete(payload.ToolInput, "tool_use_id")
-	}
-	return payload, nil
-}
-
-func parseAntigravityHookPayload(data []byte) (antigravityHookPayload, error) {
-	var raw map[string]any
-	if err := json.Unmarshal(data, &raw); err != nil {
-		return antigravityHookPayload{}, fmt.Errorf("parse antigravity hook payload: %w", err)
-	}
-	eventName := firstString(raw, "hook_event_name", "event_name", "eventName", "event", "type")
-	payload := antigravityHookPayload{
-		EventName:         eventName,
-		CWD:               firstString(raw, "cwd", "workspace", "project_dir", "projectDir"),
-		ConversationID:    firstString(raw, "conversation_id", "conversationId", "session_id", "sessionId"),
-		TranscriptPath:    firstString(raw, "transcript_path", "transcriptPath"),
-		TerminationReason: firstString(raw, "termination_reason", "terminationReason"),
-		Error:             antigravityErrorString(raw["error"]),
-		AgentState:        firstString(raw, "agent_state", "agentState"),
-		ContextWindow:     stringFromAny(firstAny(raw, "context_window", "contextWindow")),
-	}
-	if payload.CWD == "" {
-		payload.CWD = firstNestedString(raw["workspace"], "cwd", "path")
-	}
-	if payload.TranscriptPath == "" {
-		payload.TranscriptPath = firstNestedString(raw["transcript"], "path", "transcriptPath")
-	}
-	statusline := firstAny(raw, "statusline", "status_line", "statusLine", "status")
-	if payload.AgentState == "" {
-		payload.AgentState = firstNestedString(statusline, "agent_state", "agentState")
-	}
-	if payload.ContextWindow == "" {
-		payload.ContextWindow = stringFromAny(firstNestedAny(statusline, "context_window", "contextWindow"))
-	}
-	if value, ok := firstBool(raw, "fully_idle", "fullyIdle"); ok {
-		payload.FullyIdle = value
-		payload.FullyIdleSet = true
-	}
-	if value, ok := firstBool(raw, "tool_confirmation_pending", "toolConfirmationPending"); ok {
-		payload.ToolConfirmationPending = value
-		payload.ToolConfirmationPendingSet = true
-	}
-	if value, ok := firstNestedBool(statusline, "tool_confirmation_pending", "toolConfirmationPending"); ok {
-		payload.ToolConfirmationPending = value
-		payload.ToolConfirmationPendingSet = true
-	}
-	payload.EventName = normalizeAntigravityEventName(payload.EventName)
-	if payload.EventName == "" && antigravityStatuslineShapedPayload(statusline, payload) {
-		payload.EventName = "Statusline"
-	}
-	if payload.EventName == "" {
-		payload.EventName = "Unknown"
-	}
-	return payload, nil
-}
-
-func antigravityStatuslineShapedPayload(statusline any, payload antigravityHookPayload) bool {
-	return statusline != nil ||
-		strings.TrimSpace(payload.AgentState) != "" ||
-		strings.TrimSpace(payload.ContextWindow) != "" ||
-		payload.ToolConfirmationPendingSet
-}
-
-func normalizeAntigravityEventName(name string) string {
-	trimmed := strings.TrimSpace(name)
-	switch strings.ToLower(strings.ReplaceAll(trimmed, "_", "")) {
-	case "postinvocation":
-		return "PostInvocation"
-	case "stop":
-		return "Stop"
-	case "statusline", "status":
-		return "Statusline"
-	default:
-		return trimmed
-	}
-}
-
-func (p codexHookPayload) codexHookMetadata() map[string]string {
-	metadata := map[string]string{
-		"agent":           aiModeCodex,
-		"event":           p.EventName,
-		"session_id":      p.SessionID,
-		"thread_id":       p.matchThreadID(),
-		"turn_id":         p.TurnID,
-		"cwd":             p.CWD,
-		"transcript_path": p.TranscriptPath,
-		"model":           p.Model,
-		"tool_name":       p.ToolName,
-	}
-	for key, value := range p.ToolInput {
-		if text := stringFromAny(value); text != "" {
-			metadata["tool_input."+key] = truncateRunes(text, 160)
-		}
-	}
-	out := make(map[string]string, len(metadata))
-	for k, v := range metadata {
-		if value := strings.TrimSpace(v); value != "" {
-			out[k] = value
-		}
-	}
-	return out
-}
-
-func (p codexHookPayload) codexGenericHookMetadata() map[string]string {
-	metadata := map[string]string{
-		"provider":   aiHookProviderCodex,
-		"agent":      aiModeCodex,
-		"event":      p.EventName,
-		"tool":       p.ToolName,
-		"tool_name":  p.ToolName,
-		"cwd":        p.CWD,
-		"thread_id":  p.matchThreadID(),
-		"session_id": p.SessionID,
-		"turn_id":    p.TurnID,
-		"model":      p.Model,
-	}
-	out := make(map[string]string, len(metadata))
-	for k, v := range metadata {
-		if value := strings.TrimSpace(v); value != "" {
-			out[k] = value
-		}
-	}
-	return out
-}
-
-func (p codexHookPayload) matchThreadID() string {
-	if threadID := strings.TrimSpace(p.ThreadID); threadID != "" {
-		return threadID
-	}
-	return p.SessionID
-}
-
-func (p claudeHookPayload) claudeMetadata() map[string]string {
-	metadata := map[string]string{
-		"agent":             aiModeClaude,
-		"event":             p.EventName,
-		"session_id":        p.SessionID,
-		"cwd":               p.CWD,
-		"transcript_path":   p.TranscriptPath,
-		"notification_type": p.NotificationType,
-		"prompt":            truncateRunes(p.Prompt, 60),
-		"tool_name":         p.ToolName,
-		"tool_use_id":       p.ToolUseID,
-		"error_type":        p.ErrorType,
-		"error_message":     truncateRunes(p.ErrorMessage, 160),
-		"subagent_type":     p.SubagentType,
-		"subagent_id":       p.SubagentID,
-		"teammate_name":     p.TeammateName,
-		"teammate_id":       p.TeammateID,
-		"teammate_context":  truncateRunes(p.TeammateContext, 160),
-	}
-	for key, value := range p.ToolInput {
-		if text := stringFromAny(value); text != "" {
-			metadata["tool_input."+key] = truncateRunes(text, 160)
-		}
-	}
-	out := make(map[string]string, len(metadata))
-	for k, v := range metadata {
-		if value := strings.TrimSpace(v); value != "" {
-			out[k] = value
-		}
-	}
-	return out
-}
-
-func (p antigravityHookPayload) antigravityMetadata() map[string]string {
-	metadata := map[string]string{
-		"agent":              aiModeAntigravity,
-		"event":              p.EventName,
-		"conversation_id":    p.ConversationID,
-		"cwd":                p.CWD,
-		"transcript_path":    p.TranscriptPath,
-		"termination_reason": p.TerminationReason,
-		"error":              truncateRunes(p.Error, 160),
-		"agent_state":        p.AgentState,
-		"context_window":     p.ContextWindow,
-	}
-	if p.FullyIdleSet {
-		metadata["fully_idle"] = boolMetadataValue(p.FullyIdle)
-	}
-	if p.ToolConfirmationPendingSet {
-		metadata["tool_confirmation_pending"] = boolMetadataValue(p.ToolConfirmationPending)
-	}
-	out := make(map[string]string, len(metadata))
-	for k, v := range metadata {
-		if value := strings.TrimSpace(v); value != "" {
-			out[k] = value
-		}
-	}
-	return out
-}
-
 func (c *aiCommand) markAIHookPane(paneID, agent, cwd, threadID, sessionID, transcriptPath string) {
 	_ = c.run("tmux", "set-option", "-p", "-t", paneID, aiPaneHookActiveOption, "1")
 	_ = c.run("tmux", "set-option", "-p", "-t", paneID, aiPaneManagedOption, "1")
@@ -1452,155 +562,6 @@ func cleanMatchPath(path string) string {
 	return filepath.Clean(path)
 }
 
-func codexHookNotifyID(p codexHookPayload, kind string) string {
-	parts := []string{"ai", "codex", kind}
-	if value := strings.TrimSpace(p.SessionID); value != "" {
-		parts = append(parts, value)
-	}
-	if value := strings.TrimSpace(p.TurnID); value != "" {
-		parts = append(parts, value)
-	}
-	if kind == "permission" {
-		if value := strings.TrimSpace(p.ToolName); value != "" {
-			parts = append(parts, value)
-		}
-		if summary := formatCodexToolInputSummary(p.ToolName, p.ToolInput); summary != "" {
-			parts = append(parts, truncateRunes(summary, 40))
-		}
-	}
-	return strings.Join(parts, ":")
-}
-
-func claudeNotifyID(p claudeHookPayload) string {
-	parts := []string{"ai", "claude", "notification"}
-	if value := strings.TrimSpace(p.SessionID); value != "" {
-		parts = append(parts, value)
-	}
-	if value := strings.TrimSpace(p.NotificationType); value != "" {
-		parts = append(parts, value)
-	}
-	if value := strings.TrimSpace(p.Message); value != "" {
-		parts = append(parts, truncateRunes(value, 40))
-	}
-	return strings.Join(parts, ":")
-}
-
-func claudePermissionNotifyID(p claudeHookPayload) string {
-	sessionID := strings.TrimSpace(p.SessionID)
-	toolUseID := strings.TrimSpace(p.ToolUseID)
-	switch {
-	case sessionID != "" && toolUseID != "":
-		return "ai:claude:permission:" + sessionID + ":" + toolUseID
-	case toolUseID != "":
-		return "ai:claude:permission:" + toolUseID
-	default:
-		return ""
-	}
-}
-
-func claudeStopNotifyID(p claudeHookPayload) string {
-	if sessionID := strings.TrimSpace(p.SessionID); sessionID != "" {
-		return "ai:claude:stop:" + sessionID
-	}
-	return ""
-}
-
-func antigravityNotifyID(p antigravityHookPayload, kind string) string {
-	parts := []string{"ai", "antigravity", kind}
-	if value := strings.TrimSpace(p.ConversationID); value != "" {
-		parts = append(parts, value)
-	}
-	if kind == "approval" {
-		if value := strings.TrimSpace(p.AgentState); value != "" {
-			parts = append(parts, value)
-		}
-	}
-	return strings.Join(parts, ":")
-}
-
-func claudeExtraNotifyID(p claudeHookPayload, kind string, values ...string) string {
-	parts := []string{"ai", "claude", kind}
-	if value := strings.TrimSpace(p.SessionID); value != "" {
-		parts = append(parts, value)
-	}
-	for _, value := range values {
-		if trimmed := truncateRunes(value, 40); trimmed != "" {
-			parts = append(parts, trimmed)
-		}
-	}
-	return strings.Join(parts, ":")
-}
-
-func readClaudeTranscriptLastAssistantText(path string) string {
-	path = strings.TrimSpace(path)
-	if path == "" {
-		return ""
-	}
-	file, err := os.Open(path)
-	if err != nil {
-		return ""
-	}
-	defer file.Close()
-
-	info, err := file.Stat()
-	if err != nil {
-		return ""
-	}
-	size := info.Size()
-	if size <= 0 {
-		return ""
-	}
-	start := int64(0)
-	if size > claudeTranscriptTailLimit {
-		start = size - claudeTranscriptTailLimit
-	}
-	buf := make([]byte, size-start)
-	if _, err := file.ReadAt(buf, start); err != nil && !errors.Is(err, io.EOF) {
-		return ""
-	}
-	lines := strings.Split(strings.TrimSpace(string(buf)), "\n")
-	for i := len(lines) - 1; i >= 0; i-- {
-		if text := claudeAssistantTextFromJSONLine(lines[i]); text != "" {
-			return text
-		}
-	}
-	return ""
-}
-
-func claudeAssistantTextFromJSONLine(line string) string {
-	var raw map[string]any
-	if err := json.Unmarshal([]byte(line), &raw); err != nil {
-		return ""
-	}
-	if strings.EqualFold(stringFromAny(raw["role"]), "assistant") {
-		return claudeContentText(raw["content"])
-	}
-	message := mapFromAny(raw["message"])
-	if strings.EqualFold(stringFromAny(message["role"]), "assistant") {
-		return claudeContentText(message["content"])
-	}
-	return ""
-}
-
-func claudeContentText(value any) string {
-	switch v := value.(type) {
-	case string:
-		return strings.TrimSpace(v)
-	case []any:
-		for _, item := range v {
-			if text := firstNestedString(item, "text"); text != "" {
-				return text
-			}
-			if text := stringFromAny(item); text != "" {
-				return text
-			}
-		}
-	case map[string]any:
-		return firstString(v, "text", "content")
-	}
-	return ""
-}
-
 func firstString(raw map[string]any, keys ...string) string {
 	for _, key := range keys {
 		if value := stringFromAny(raw[key]); value != "" {
@@ -1659,20 +620,6 @@ func mapFromAny(value any) map[string]any {
 		return out
 	}
 	return map[string]any{}
-}
-
-func antigravityErrorString(value any) string {
-	switch v := value.(type) {
-	case nil:
-		return ""
-	case string:
-		return strings.TrimSpace(v)
-	case map[string]any:
-		if text := firstString(v, "message", "text", "reason", "type", "code"); text != "" {
-			return text
-		}
-	}
-	return stringFromAny(value)
 }
 
 func boolFromAny(value any) (bool, bool) {
